Reorganise navdata to it's own directory
