+++ frontend/src/core/App.ts
@@ -22,9 +22,9 @@ import { MapDisplay } from '../ui/map/MapDisplay';
 import { MapOverlay } from '../ui/map/MapOverlay';
 import { ShapeDrawingManager } from '../ui/map/shapes/ShapeDrawingManager';
 import { ShapeRenderer } from '../ui/map/shapes/ShapeRenderer';
-import { NavdataRenderer } from '../ui/map/NavdataRenderer';
-import { NavaidSnapper } from '../ui/map/NavaidSnapper';
-import { NavSearchBox } from '../ui/map/NavSearchBox';
+import { NavdataRenderer } from '../ui/map/navdata/NavdataRenderer';
+import { NavaidSnapper } from '../ui/map/navdata/NavaidSnapper';
+import { NavSearchBox } from '../ui/map/navdata/NavSearchBox';
 import { AircraftCreationManager } from '../ui/map/aircraft/AircraftCreationManager';
 import { AircraftInteractionManager } from '../ui/map/aircraft/AircraftInteractionManager';
 import { RouteDrawingManager } from '../ui/map/routes/RouteDrawingManager';
